fix: remove logtail logging (#106)
Currently pushing 8GB of logs per day. The contents of which are available in Cloudflare already (we do not log any further info). Errors are pushed to Sentry.

Also adds `multiformats` to the client dependencies which was resolved in https://github.com/storacha/w3name/pull/103 but was not mergable because it did not also update the lockfile.

---------

Co-authored-by: Duncan <legowerewolf@gmail.com>